refactor: move batchnormed layers to their base versions
